split cmd and api into separate files
also fix some lint issues, fix godoc style, replace Retrieve function w/ inline vault api call, and rename wrapItUp
